requirements.txt
# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.52.0  # https://github.com/pennersr/django-allauth
djangorestframework==3.14.0
djangorestframework_simplejwt==4.7.0
django-filter==22.1
django-url-filter==0.3.15
django-cors-headers==3.13.0
# ------------------------------------------------------------------------------
ipdb==0.13.11  # https://github.com/gotcha/ipdb
#psycopg2==2.9.5  # https://github.com/psycopg/psycopg2
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
psycopg2-binary==2.9.5
pytest-dotenv>=0.5.2<0.6
python-dotenv>=1.0.0<1.1
drf-yasg==1.21.5
pyjwt>=1.7.1
requests>=2.31.0
drf-access-policy>=1.5.0,<1.6
#html5lib==1.0.1
#PyPDF2==1.26.0
#pytz==2020.1
#reportlab==3.5.42
#six==1.14.0
#sqlparse==0.3.1
#Unipath==1.1
#webencodings==0.5.1
weasyprint==62.3
cryptography==41.0.3
Babel==2.15.0
num2words==0.5.13
#mozilla-django-oidc==3.0.0

openpyxl>=3.1.2